login function
This is the login function without the password protect system.
fgetc() or getchar() and fputc() or putchar() can be used to create a system to create astericks but the issue comes with the echo from the keyboard still being there. It also prevents the loop from completing properly and issuing a message confirming or deying login